requirements.txt
airtest==1.1.3
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
comtypes==1.1.7
cryptography==2.9.2
facebook-wda==0.2.1
hrpc==1.0.8
idna==2.9
Jinja2==2.11.2
MarkupSafe==1.1.1
mss==4.0.3
numpy==1.18.4
opencv-contrib-python==3.4.2.17
opencv-python==4.3.0.36
pandas==1.0.5
Pillow==7.1.2
pocoui==1.0.79
pocounit==1.0.40
pycparser==2.20
PyJWT==1.7.1
pypiwin32==223
python-dateutil==2.8.1
pytz==2020.1
pywin32==227
pywinauto==0.6.3
requests==2.24.0
six==1.15.0
urllib3==1.25.9
websocket-client==0.48.0
xlrd==1.2.0
